+++ src/app/api/v1/api-keys/[access_key_id]/api-keys.access_key_id.test.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest } from "next";
 import httpMocks from "node-mocks-http";
 import { handler } from "@/pages/api/v1/api-keys/[access_key_id]";
-import { getSession } from "@/api/utils";
+import { getServerSession } from "@ory/nextjs/app";
 import { isAuthorized } from "@/api/authz";
 import { getAPIKey, putAPIKey } from "@/api/db";
 import { UnauthorizedError, MethodNotImplementedError } from "@/api/errors";
@@ -9,7 +9,7 @@ import { MockNextApiResponse, jsonBody } from "@/api/utils/mock";
 import { AccountType, UserSession, APIKey, Actions } from "@/api/types";
 
 jest.mock("@/api/utils", () => ({
-  getSession: jest.fn(),
+  getServerSession: jest.fn(),
 }));
 
 jest.mock("@/api/authz", () => ({
@@ -41,7 +41,9 @@ describe("/api/v1/api-keys/[access_key_id]", () => {
     });
 
     it("should throw UnauthorizedError when API key is not found", async () => {
-      (getSession as jest.Mock).mockResolvedValue({ identity_id: "user-1" });
+      (getServerSession as jest.Mock).mockResolvedValue({
+        identity_id: "user-1",
+      });
       (getAPIKey as jest.Mock).mockResolvedValue(null);
 
       await expect(handler(req, res)).rejects.toThrow(UnauthorizedError);
@@ -57,7 +59,7 @@ describe("/api/v1/api-keys/[access_key_id]", () => {
         secret_access_key:
           "A1b2C3d4E5f6G7h8I9j0K1l2M3n4O5p6Q7r8S9t0U1v2W3x4Y5z6A7B8C9D00000",
       };
-      (getSession as jest.Mock).mockResolvedValue({
+      (getServerSession as jest.Mock).mockResolvedValue({
         identity_id: "unauthorized-user",
       });
       (getAPIKey as jest.Mock).mockResolvedValue(mockAPIKey);
@@ -77,7 +79,7 @@ describe("/api/v1/api-keys/[access_key_id]", () => {
           "A1b2C3d4E5f6G7h8I9j0K1l2M3n4O5p6Q7r8S9t0U1v2W3x4Y5z6A7B8C9D00000",
       };
       const revokedAPIKey: APIKey = { ...mockAPIKey, disabled: true };
-      (getSession as jest.Mock).mockResolvedValue({
+      (getServerSession as jest.Mock).mockResolvedValue({
         identity_id: "authorized-user",
       });
       (getAPIKey as jest.Mock).mockResolvedValue(mockAPIKey);
