+++ src/app/api/v1/memberships/[membership_id]/accept.test.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest } from "next";
 import httpMocks from "node-mocks-http";
 import { handler } from "@/pages/api/v1/memberships/[membership_id]/accept";
-import { getSession } from "@/api/utils";
+import { getServerSession } from "@ory/nextjs/app";
 import { isAuthorized } from "@/api/authz";
 import { getMembership, putMembership } from "@/api/db";
 import {
@@ -20,7 +20,7 @@ import {
 import logger from "@/utils/logger";
 
 jest.mock("@/api/utils", () => ({
-  getSession: jest.fn(),
+  getServerSession: jest.fn(),
 }));
 
 jest.mock("@/api/authz", () => ({
@@ -48,7 +48,9 @@ describe("/api/v1/memberships/[membership_id]/accept", () => {
   });
 
   it("should throw NotFoundError when membership doesn't exist", async () => {
-    (getSession as jest.Mock).mockResolvedValue({ identity_id: "user-1" });
+    (getServerSession as jest.Mock).mockResolvedValue({
+      identity_id: "user-1",
+    });
     (getMembership as jest.Mock).mockResolvedValue(null);
 
     await expect(handler(req, res)).rejects.toThrow(NotFoundError);
@@ -63,7 +65,7 @@ describe("/api/v1/memberships/[membership_id]/accept", () => {
       state: MembershipState.Invited,
       state_changed: new Date().toISOString(),
     };
-    (getSession as jest.Mock).mockResolvedValue({
+    (getServerSession as jest.Mock).mockResolvedValue({
       identity_id: "unauthorized-user",
     });
     (getMembership as jest.Mock).mockResolvedValue(mockMembership);
@@ -81,7 +83,7 @@ describe("/api/v1/memberships/[membership_id]/accept", () => {
       state: MembershipState.Member,
       state_changed: new Date().toISOString(),
     };
-    (getSession as jest.Mock).mockResolvedValue({
+    (getServerSession as jest.Mock).mockResolvedValue({
       identity_id: "authorized-user",
     });
     (getMembership as jest.Mock).mockResolvedValue(mockMembership);
@@ -99,7 +101,7 @@ describe("/api/v1/memberships/[membership_id]/accept", () => {
       state: MembershipState.Revoked,
       state_changed: new Date().toISOString(),
     };
-    (getSession as jest.Mock).mockResolvedValue({
+    (getServerSession as jest.Mock).mockResolvedValue({
       identity_id: "authorized-user",
     });
     (getMembership as jest.Mock).mockResolvedValue(mockMembership);
@@ -117,7 +119,7 @@ describe("/api/v1/memberships/[membership_id]/accept", () => {
       state: MembershipState.Invited,
       state_changed: "2023-01-01T00:00:00Z", // Use a fixed date for testing
     };
-    (getSession as jest.Mock).mockResolvedValue({
+    (getServerSession as jest.Mock).mockResolvedValue({
       identity_id: "authorized-user",
     });
     (getMembership as jest.Mock).mockResolvedValue(mockMembership);
