+++ src/app/api/v1/repositories/[account_id]/[repository_id]/api-keys/api-keys.test.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest } from "next";
 import httpMocks from "node-mocks-http";
 import { handler } from "@/pages/api/v1/repositories/[account_id]/[repository_id]/api-keys";
-import { getSession } from "@/api/utils";
+import { getServerSession } from "@ory/nextjs/app";
 import { isAuthorized } from "@/api/authz";
 import { getRepository, putAPIKey, getAPIKeys } from "@/api/db";
 import {
@@ -25,7 +25,7 @@ import {
 } from "@/api/types";
 
 jest.mock("@/api/utils", () => ({
-  getSession: jest.fn(),
+  getServerSession: jest.fn(),
   generateAccessKeyID: jest.fn(() => "SCgenerated-access-key-id"),
   generateSecretAccessKey: jest.fn(
     () => "A1b2C3d4E5f6G7h8I9j0K1l2M3n4O5p6Q7r8S9t0U1v2W3x4Y5z6A7B8C9D00000"
@@ -66,7 +66,9 @@ describe("/api/v1/repositories/[account_id]/[repository_id]/api-keys", () => {
     });
 
     it("should throw NotFoundError when repository doesn't exist", async () => {
-      (getSession as jest.Mock).mockResolvedValue({ identity_id: "user-1" });
+      (getServerSession as jest.Mock).mockResolvedValue({
+        identity_id: "user-1",
+      });
       (getRepository as jest.Mock).mockResolvedValue(null);
 
       await expect(handler(req, res)).rejects.toThrow(NotFoundError);
@@ -74,7 +76,9 @@ describe("/api/v1/repositories/[account_id]/[repository_id]/api-keys", () => {
 
     it("should throw BadRequestError when expiration date is in the past", async () => {
       req.body.expires = new Date(Date.now() - 86400000).toISOString(); // 1 day in the past
-      (getSession as jest.Mock).mockResolvedValue({ identity_id: "user-1" });
+      (getServerSession as jest.Mock).mockResolvedValue({
+        identity_id: "user-1",
+      });
       (getRepository as jest.Mock).mockResolvedValue({
         account_id: "test-account",
         repository_id: "test-repo",
@@ -102,7 +106,7 @@ describe("/api/v1/repositories/[account_id]/[repository_id]/api-keys", () => {
         published: new Date().toISOString(),
         disabled: false,
       };
-      (getSession as jest.Mock).mockResolvedValue({
+      (getServerSession as jest.Mock).mockResolvedValue({
         identity_id: "unauthorized-user",
       });
       (getRepository as jest.Mock).mockResolvedValue(mockRepository);
@@ -139,7 +143,7 @@ describe("/api/v1/repositories/[account_id]/[repository_id]/api-keys", () => {
         secret_access_key:
           "A1b2C3d4E5f6G7h8I9j0K1l2M3n4O5p6Q7r8S9t0U1v2W3x4Y5z6A7B8C9D00000",
       };
-      (getSession as jest.Mock).mockResolvedValue({
+      (getServerSession as jest.Mock).mockResolvedValue({
         identity_id: "authorized-user",
       });
       (getRepository as jest.Mock).mockResolvedValue(mockRepository);
@@ -168,7 +172,9 @@ describe("/api/v1/repositories/[account_id]/[repository_id]/api-keys", () => {
     });
 
     it("should throw NotFoundError when repository doesn't exist", async () => {
-      (getSession as jest.Mock).mockResolvedValue({ identity_id: "user-1" });
+      (getServerSession as jest.Mock).mockResolvedValue({
+        identity_id: "user-1",
+      });
       (getRepository as jest.Mock).mockResolvedValue(null);
 
       await expect(handler(req, res)).rejects.toThrow(NotFoundError);
@@ -193,7 +199,7 @@ describe("/api/v1/repositories/[account_id]/[repository_id]/api-keys", () => {
         published: new Date().toISOString(),
         disabled: false,
       };
-      (getSession as jest.Mock).mockResolvedValue({
+      (getServerSession as jest.Mock).mockResolvedValue({
         identity_id: "unauthorized-user",
       });
       (getRepository as jest.Mock).mockResolvedValue(mockRepository);
@@ -243,7 +249,7 @@ describe("/api/v1/repositories/[account_id]/[repository_id]/api-keys", () => {
             "A1b2C3d4E5f6G7h8I9j0K1l2M3n4O5p6Q7r8S9t0U1v2W3x4Y5z6A7B8C9D00000",
         },
       ];
-      (getSession as jest.Mock).mockResolvedValue({
+      (getServerSession as jest.Mock).mockResolvedValue({
         identity_id: "authorized-user",
       });
       (getRepository as jest.Mock).mockResolvedValue(mockRepository);
